mv factory for graphics overlay collection
