+++ toolkit/geo-compose/src/main/java/com/arcgismaps/toolkit/geocompose/GraphicsOverlayCollection.kt
@@ -20,6 +20,7 @@ package com.arcgismaps.toolkit.geocompose
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.Stable
+import androidx.compose.runtime.remember
 import com.arcgismaps.mapping.view.GeoView
 import com.arcgismaps.mapping.view.GraphicsOverlay
 import kotlinx.coroutines.channels.BufferOverflow
@@ -143,4 +144,21 @@ internal fun GraphicsOverlaysUpdater(
             }
         }
     }
+}
+
+/**
+ * Create and [remember] a [GraphicsOverlayCollection].
+ * [init] will be called when the [GraphicsOverlayCollection] is first created to configure its
+ * initial state.
+ *
+ * @param key invalidates the remembered GraphicsOverlayCollection if different from the previous composition
+ * @param init called when the [GraphicsOverlayCollection] is created to configure its initial state
+ * @since 200.4.0
+ */
+@Composable
+public inline fun rememberGraphicsOverlayCollection(
+    key: Any? = null,
+    crossinline init: GraphicsOverlayCollection.() -> Unit = {}
+): GraphicsOverlayCollection = remember(key) {
+    GraphicsOverlayCollection().apply(init)
 }
